Remove more implicit casts in upnp part
Remove more "implicit integer or enum conversions" errors as well as
dead code.

(forward port of commit c40d2bc0c9b60c43b641ac4669c7b8bbcd6134c5)

+++ upnp/src/inc/config.h
@@ -126,7 +126,7 @@
  *
  * @{
  */
-#define THREAD_STACK_SIZE 0
+#define THREAD_STACK_SIZE (size_t)0
 /* @} */
